refacto handle contacts file
